backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
alpaca-py==0.38.0
annotated-types==0.6.0
anthropic==0.47.2
anyio==4.8.0
attrs>=23.0
bidict==0.22.1
blinker==1.7.0
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
coverage==7.6.12
cryptography>=43.0.3,<44.0.0
dataclasses-json==0.6.7
distro==1.9.0
dydantic==0.0.8
exceptiongroup==1.2.0
fastapi>=0.109.3
filelock==3.16.1
flake8==7.1.2
frozenlist==1.5.0
fsspec==2024.12.0
groq==0.18.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.1
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
langchain>=0.3.9
langchain-anthropic>=0.1.23
langchain-community>=0.2.17 # 0.3.18
langchain-core>=0.2.27
langchain-groq>=0.2.3
langchain-huggingface>=0.1.2
langchain-openai>=0.1.22
langchain-pinecone>=0.2.2
langchain-tests>=0.3.8
langchain-text-splitters>=0.3.6
langgraph>=0.2.73
langgraph-api>=0.0.27
langgraph-checkpoint>=2.0.16
langgraph-cli>=0.1.74
langgraph-sdk>=0.1.51
langgraph-supervisor>=0.0.2
langsmith>=0.3.11
MarkupSafe==2.1.4
marshmallow==3.25.1
mccabe==0.7.0
mock==5.1.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
ngrok==1.4.0
nose==1.3.7
numpy==1.26.4
openai==1.64.0
orjson==3.10.14
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
pip-conflict-checker==0.6.0
pipdeptree==2.25.0
pluggy==1.5.0
propcache==0.2.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyflakes==3.2.0
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv>=1.0.1
python-engineio==4.8.2
python-multipart==0.0.9
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
retell==0.5.1
retell-sdk==4.20.0
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.3.1
setuptools==75.8.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.0
sounddevice==0.5.1
soundfile==0.13.1
SQLAlchemy==2.0.37
sseclient-py==1.8.0
starlette>=0.38.6
sympy==1.13.1
syrupy==4.8.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.66.1
transformers==4.48.0
trustcall==0.0.37
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
websockets==14.2
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.18.3
zipp==3.17.0
zstandard==0.23.0
livekit-agents>=0.12.11,<1.0.0
livekit-plugins-openai>=0.10.17,<1.0.0
livekit-plugins-cartesia>=0.4.7,<1.0.0
livekit-plugins-deepgram>=0.6.17,<1.0.0
livekit-plugins-silero>=0.7.4,<1.0.0
livekit-plugins-turn-detector>=0.4.0,<1.0.0